Made AddCommand and SubtractCommand use Long instead of Integer.

+++ src/main/java/com/haxwell/disposableIncomeScheduler/report/commands/AddCommand.java
@@ -2,22 +2,22 @@ package com.haxwell.disposableIncomeScheduler.report.commands;
 
 public class AddCommand extends Command {
 	
-	protected int amt = -1;
+	protected long amt = -1;
 	protected String grp = null;
 	protected String desc = null;
 	
-	public AddCommand(int amount, String desc) {
+	public AddCommand(long amount, String desc) {
 		this.amt = amount;
 		this.desc = desc; 
 	}
 
-	public AddCommand(int amount, String desc, String group) {
+	public AddCommand(long amount, String desc, String group) {
 		this.amt = amount;
 		this.grp = group;
 		this.desc = desc;
 	}
 	
-	public int getAmount() {
+	public long getAmount() {
 		return amt;
 	}
